requirements.txt
pytube==15.0.0
requests==2.31.0
easyocr==1.7.1
paddleocr==2.7.0.3
python-dateutil==2.8.2
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
paddlepaddle==2.5.1
python-dotenv==1.0.0
discord.py==2.3.2
